Group helper for list segment boundaries.

Given the segment positions, the index and the total time, returns a list of segment boundary positions and intervals.

Usage

list_segment_boundaries_group_helper(.positions, .index, .total)

Arguments

.positions

Segment positions.

.index

Data index.

.total

Data total.

Value

Tibble with segment boundary list.

Details

The computation assumes that a segment boundary exists in either boundary repetition, the boundary interval is therefore formed by the mid points of the two repetitions, arguing that less than half of each boundary repetition is impacted (otherwise segments would be assigned differently).

Head and tail boundaries are not included.