Load data from given file.

Load data from given file.

Usage

load_file_json(data_file)

Arguments

data_file

Data file to load from.

Value

Tibble with loaded data.

Details

The basic timing columns are translated into canonical form:

time will contain individual repetition time in seconds

total will contain aggregate execution time at the end of repetition in seconds

index will contain the index of the measurement in the run starting from 1

Additional metadata is included as factor columns:

vm is a hash of all configuration metadata and can be used to distinguish incompatible measurement configurations

vm_name is the name of the virtual machine

vm_version is the version of the virtual machine

vm_configuration gives concatenated command line arguments

run is a unique identifier of the run

benchmark is the name of the benchmark

See also

Examples

load_file_json (rren_example ('results-small-version-5.json'))
#> INFO [2025-02-12 11:58:07] Loaded 6 rows from file /home/runner/work/_temp/Library/rren/extdata/results-small-version-5.json.
#> # A tibble: 6 × 9
#>   index  time total vm_name    vm_version vm_configuration run   vm    benchmark
#>   <int> <dbl> <dbl> <fct>      <fct>      <fct>            <fct> <fct> <fct>    
#> 1     1  5.41  10.4 OpenJDK 6… 17.0.3+7   ""               /hom… cec0… page-rank
#> 2     2  2.45  14.9 OpenJDK 6… 17.0.3+7   ""               /hom… cec0… page-rank
#> 3     3  2.33  18.5 OpenJDK 6… 17.0.3+7   ""               /hom… cec0… page-rank
#> 4     1  7.80  30.3 OpenJDK 6… 17.0.3+7   ""               /hom… cec0… naive-ba…
#> 5     2  1.43  31.7 OpenJDK 6… 17.0.3+7   ""               /hom… cec0… naive-ba…
#> 6     3  1.28  33.0 OpenJDK 6… 17.0.3+7   ""               /hom… cec0… naive-ba…